+++ lib/src/osid_unknown.dart
@@ -2,10 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../override.dart';
+import "../override.dart";
 
-const String _os =
-    String.fromEnvironment("dart.os.name", defaultValue: "unknown");
-const String _osVersion = String.fromEnvironment("dart.os.version");
-
-const OperatingSystem platformOS = OperatingSystem(_os, _osVersion);
+const OperatingSystem platformOS = OperatingSystem("unknown", "");